requirements.txt
django==3.0.7
django-debug-toolbar==2.2
dj-database-url==0.5.0
Pillow==7.1.2
django-phonenumber-field==5.0.0
djangorestframework==3.12.2
